change macro for registering imgui for components, fix crash in TestComp, fix backspace in Imgui, add Component handling fuctions to Object

+++ Engine/Components/TestComponent.hpp
@@ -21,7 +21,7 @@ public:
 
   void FromJson(json &js) override;
 
-  int amount = 0;
+  float amount = 0;
 
 private:
 };
